Bootstrap builds just PerformanceTesting Project
As part of running the tests, whole of Roslyn will be built and so will be the csi proj. Hence bootstrap process need not build csi proj any more. Also all the tests are expected to run from the binaries folder. Hence, we need to build and move the the perf binary helpers and perf tests as part of bootstrap